Move extractZipFile to server/util
