+++ server/scripts/import-tests/testPlanVersionOperations.js
@@ -4,7 +4,6 @@ const path = require('path');
 const fse = require('fs-extra');
 const { pathToFileURL } = require('url');
 const spawn = require('cross-spawn');
-const yauzl = require('yauzl');
 const { At, sequelize } = require('../../models');
 const {
   createTestPlanVersion,
@@ -16,6 +15,7 @@ const {
   createTestPlan
 } = require('../../models/services/TestPlanService');
 const { hashTests } = require('../../util/aria');
+const extractZipFile = require('../../util/extractZipFile');
 const { dates } = require('shared');
 const { readCommit, readDirectoryGitInfo } = require('./gitOperations');
 const { parseTests } = require('./testParser');
@@ -28,52 +28,6 @@ const {
   PRE_BUILT_ZIP_COMMITS
 } = require('./settings');
 const { getAppUrl } = require('./utils');
-
-/**
- * Extracts a zip file to a target directory.
- * @param {string} zipPath - Path to the zip file.
- * @param {string} targetPath - Path to extract to.
- * @returns {Promise<void>}
- */
-const extractZipFile = (zipPath, targetPath) => {
-  return new Promise((resolve, reject) => {
-    yauzl.open(zipPath, { lazyEntries: true }, (err, zipfile) => {
-      if (err) return reject(err);
-
-      zipfile.on('close', () => resolve());
-      zipfile.on('error', reject);
-
-      zipfile.on('entry', entry => {
-        if (/\/$/.test(entry.fileName)) {
-          // Directory entry
-          zipfile.readEntry();
-        } else {
-          // File entry
-          zipfile.openReadStream(entry, (err, readStream) => {
-            if (err) return reject(err);
-
-            const filePath = path.join(targetPath, entry.fileName);
-            const dirPath = path.dirname(filePath);
-
-            // Ensure directory exists
-            fse.mkdirSync(dirPath, { recursive: true });
-
-            const writeStream = fse.createWriteStream(filePath);
-            readStream.pipe(writeStream);
-
-            writeStream.on('close', () => {
-              zipfile.readEntry();
-            });
-
-            writeStream.on('error', reject);
-          });
-        }
-      });
-
-      zipfile.readEntry();
-    });
-  });
-};
 
 /**
  * Builds tests and creates test plan versions for a given commit.
